+++ src/lib/libmalloc/malloc.c
@@ -1,7 +1,7 @@
 
 #include "_malloc.h"
 
-void* malloc(int size)
+void* malloc(size_t size)
 {
     return heman_alloc(heap_memory, size);
 }
